Plot multiple variables with plot_ensemble.py
You can now use a comma-separated list of variables to plot a new variable in each row.
There is also now a --plotChange flag. If True, it plots the change in the variable
since time 0. If False, it plots the variable as-is. Sea-level contribution axis
works with both --plotChange True and False