Add script to plot simple regional stats
Add script to plot simple regional stats. Currently this is a hacked
together script based on plot_regionalStats.py. There are lots of
hardcoded plot settings relevant to our AGU poster on ISMIP6 results
that could be generalized in the future. The user specifies up to 6
regional stats files to plot from, as well as numeric values for
which regions to plot, and the name of the variable to plot.